Updates. Not too much exciting stuff, mind you. Mostly theoretical work resulting in slightly different code.
On a sidenote, I do not think the name 'self consistent' will remain in use for long. Part of this is that this is not truly a self-consistent
equation. While in principle, we could state that P = K^{-1} W P is self consistent, we are more interested in the null-space.

Indeed, I would postulate that a system that does not result in a proper nullspace is in some sense unphysical, or at least does not fit our model.
Despite the lack of a guarantee, the math is clear and consistency requires a well defined nullspace.

Based on that choice, I have made some theoretical decisions.