Results.
Final calculation still has to be done. But the number occupation expectations seem to be both symmetric and changing as teh temperature does, which is exactly what should be expected.

As for the transport contour, it seems to include both state switching and thermal broadening, exactly as it should be. Now, let's hope we can establish the Kondo Effect in this model.